+++ extras/Mesa/src/mesa/drivers/dri/i915/intel_batchbuffer.c
@@ -1,3 +1,4 @@
+/* $XFree86$ */
 /**************************************************************************
  * 
  * Copyright 2003 Tungsten Graphics, Inc., Cedar Park, Texas.
@@ -188,7 +189,7 @@ void intelStartInlinePrimitive( intelContextPtr intel, GLuint prim )
 
 
 #if 1
-   if (((int)intel->batch.ptr) & 0x4) {
+   if (((unsigned long)intel->batch.ptr) & 0x4) {
       BEGIN_BATCH(1);
       OUT_BATCH(0);
       ADVANCE_BATCH();
